+++ app/middleware/staticFileMiddleware.ts
@@ -15,6 +15,7 @@ export const staticFileMiddleware = async (ctx: Context, next: Function) => {
 async function fileExists(path: string) {
   try {
     const stats = await Deno.lstat(path);
+    
     return stats && stats.isFile;
   } catch (e) {
     if (e && e instanceof Deno.errors.NotFound) {